week2-image-audio/class07-audio-mcp/requirements.txt
google-genai
librosa
soundfile
sounddevice
numpy
pynput
pygame
python-dotenv